+++ src/parlant/core/engines/alpha/message_event_generator.py
@@ -138,7 +138,7 @@ class MessageEventGenerator:
                 return []
 
             self._logger.debug(
-                f"""[MessageEventGenerator] Guidelines applied: {
+                f"""[MessageEventGenerator] Guidelines applied\n{
                     json.dumps(
                         [
                             {
@@ -188,7 +188,7 @@ class MessageEventGenerator:
 
             last_generation_exception: Exception | None = None
 
-            self._logger.debug(f"[MessageEventGenerator][Prompt] \n{prompt}")
+            self._logger.debug(f"[MessageEventGenerator][Prompt]\n{prompt}")
 
             for generation_attempt in range(3):
                 try:
@@ -530,17 +530,6 @@ Produce a valid JSON object in the following format: ###
         if not message_event_response.content.produced_reply:
             self._logger.debug("[MessageEventGenerator] Produced no reply")
             return message_event_response.info, None
-
-        if message_event_response.content.evaluation_for_each_instruction:
-            self._logger.debug(
-                "[MessageEventGenerator][Evaluations]\n"
-                f"{json.dumps([e.model_dump(mode='json') for e in message_event_response.content.evaluation_for_each_instruction], indent=2)}"
-            )
-
-        self._logger.debug(
-            "[MessageEventGenerator][Revisions]\n"
-            f"{json.dumps([r.model_dump(mode='json') for r in message_event_response.content.revisions], indent=2)}"
-        )
 
         if first_correct_revision := next(
             (